Politicas de acceso 06_01

+++ app/Http/Controllers/UsersController.php
@@ -13,8 +13,8 @@ class UsersController extends Controller
 
     function __construct()
     {
-        $this->middleware('auth');
-        $this->middleware('roles:admin', ['except' => ['edit']]);
+        $this->middleware('auth', ['except' => ['show']]);
+        $this->middleware('roles:admin', ['except' => ['edit', 'update', 'show']]);
     }
 
     /**
@@ -58,7 +58,9 @@ class UsersController extends Controller
      */
     public function show($id)
     {
-        //
+        $user = User::findOrFail($id);
+        
+        return view('users.show', compact('user'));
     }
 
     /**
@@ -70,6 +72,7 @@ class UsersController extends Controller
     public function edit($id)
     {
         $user = User::findOrFail($id);
+        $this->authorize('edit',$user);
         return view('users.edit', compact('user'));
     }
 
@@ -83,6 +86,7 @@ class UsersController extends Controller
     public function update(UpdateUserRequest $request, $id)
     {
         $user = User::findOrFail($id);
+        $this->authorize('update',$user);
         $user->update($request->all());
         return back()->with('info', 'Usuario actualizado');
     }
@@ -95,8 +99,9 @@ class UsersController extends Controller
      */
     public function destroy($id)
     {
+        $this->authorize('destroy',$user);
         User::findOrFail($id)->delete();
         Alert::success('Usuario eliminado!')->persistent("Cerrar");
-        return redirect()->route('usuarios.index');
+        return back();
     }
 }
